Change the way the icon font is loaded and set to make it easier
The font is now loaded once with a default size (12) and then can be set using the icon css class, which also removes extra padding. This means it can be set from Scene Builder rather than having to be done programmatically which was tedious. (The text still has to be set programmatically.) The font can also be set using the getIconFont method if it makes more sense to be set programmatically. This method does not remove extra padding however.

Updated everwhere the icons are used to use the new way.